move CompatibleMode into config
