+++ docs/guide/java_object_graph_guide.md
@@ -332,7 +332,7 @@ if(JavaSerializer.serializedByJDK(bytes)){
 
 ## Trouble shooting
 ### Class inconsistency and class version check
-If you create fury without setting `CompatibleMode` to `io.fury.serializer.CompatibleMode.COMPATIBLE`, and you got a strange
+If you create fury without setting `CompatibleMode` to `io.fury.config.CompatibleMode.COMPATIBLE`, and you got a strange
 serialization error, it may be caused by class inconsistency between serialization peer and deserialization peer.
 
 In such cases, you can invoke `FuryBuilder#withClassVersionCheck` to create fury to validate it, if deserialization throws `io.fury.exception.ClassNotCompatibleException`, it shows class are inconsistent, and you should create fury with
