+++ integration_tests/latest_jdk_tests/src/test/java/io/fury/integration_tests/RecordSerializersTest.java
@@ -21,7 +21,7 @@ import static io.fury.collection.Collections.ofHashMap;
 
 import io.fury.Fury;
 import io.fury.resolver.MetaContext;
-import io.fury.serializer.CompatibleMode;
+import io.fury.config.CompatibleMode;
 import io.fury.test.bean.Struct;
 import io.fury.util.record.RecordComponent;
 import io.fury.util.record.RecordUtils;
