+++ java/fury-benchmark/src/main/java/io/fury/benchmark/ZeroCopySuite.java
@@ -29,10 +29,10 @@ import com.esotericsoftware.kryo.io.UnsafeMemoryInput;
 import com.esotericsoftware.kryo.io.UnsafeMemoryOutput;
 import com.google.common.base.Preconditions;
 import io.fury.Fury;
-import io.fury.config.Language;
 import io.fury.benchmark.data.ComparableByteArray;
 import io.fury.benchmark.data.SerializableByteBuffer;
 import io.fury.benchmark.state.BufferType;
+import io.fury.config.Language;
 import io.fury.memory.MemoryBuffer;
 import io.fury.memory.MemoryUtils;
 import io.fury.serializer.BufferObject;
