+++ java/fury-benchmark/src/main/java/io/fury/benchmark/state/FuryState.java
@@ -18,7 +18,6 @@ package io.fury.benchmark.state;
 
 import com.google.common.base.Preconditions;
 import io.fury.Fury;
-import io.fury.config.Language;
 import io.fury.benchmark.IntsSerializationSuite;
 import io.fury.benchmark.LongStringSerializationSuite;
 import io.fury.benchmark.LongsSerializationSuite;
@@ -30,11 +29,12 @@ import io.fury.benchmark.data.Image;
 import io.fury.benchmark.data.Media;
 import io.fury.benchmark.data.MediaContent;
 import io.fury.benchmark.data.Struct;
+import io.fury.config.CompatibleMode;
 import io.fury.config.FuryBuilder;
+import io.fury.config.Language;
 import io.fury.memory.MemoryBuffer;
 import io.fury.memory.MemoryUtils;
 import io.fury.resolver.MetaContext;
-import io.fury.serializer.CompatibleMode;
 import io.fury.util.LoggerFactory;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
