+++ java/fury-core/src/main/java/io/fury/builder/MetaSharedCodecBuilder.java
@@ -25,10 +25,10 @@ import io.fury.builder.Generated.GeneratedMetaSharedSerializer;
 import io.fury.codegen.CodeGenerator;
 import io.fury.codegen.Expression;
 import io.fury.codegen.Expression.Literal;
+import io.fury.config.CompatibleMode;
 import io.fury.config.FuryBuilder;
 import io.fury.memory.MemoryBuffer;
 import io.fury.serializer.CodegenSerializer;
-import io.fury.serializer.CompatibleMode;
 import io.fury.serializer.MetaSharedSerializer;
 import io.fury.serializer.ObjectSerializer;
 import io.fury.serializer.Serializer;
