+++ java/fury-core/src/main/java/io/fury/resolver/ClassResolver.java
@@ -43,6 +43,7 @@ import io.fury.collection.IdentityObjectIntMap;
 import io.fury.collection.LongMap;
 import io.fury.collection.ObjectMap;
 import io.fury.collection.Tuple2;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.exception.InsecureException;
 import io.fury.memory.MemoryBuffer;
@@ -51,7 +52,6 @@ import io.fury.serializer.BufferSerializers;
 import io.fury.serializer.ChildContainerSerializers;
 import io.fury.serializer.CodegenSerializer.LazyInitBeanSerializer;
 import io.fury.serializer.CollectionSerializers;
-import io.fury.serializer.CompatibleMode;
 import io.fury.serializer.CompatibleSerializer;
 import io.fury.serializer.ExternalizableSerializer;
 import io.fury.serializer.GuavaSerializers;
