+++ java/fury-core/src/main/java/io/fury/type/ClassDef.java
@@ -26,11 +26,11 @@ import com.google.common.reflect.TypeToken;
 import io.fury.Fury;
 import io.fury.builder.MetaSharedCodecBuilder;
 import io.fury.collection.IdentityObjectIntMap;
+import io.fury.config.CompatibleMode;
 import io.fury.config.FuryBuilder;
 import io.fury.memory.MemoryBuffer;
 import io.fury.memory.MemoryUtils;
 import io.fury.resolver.ClassResolver;
-import io.fury.serializer.CompatibleMode;
 import io.fury.serializer.CompatibleSerializer;
 import io.fury.util.LoggerFactory;
 import io.fury.util.MurmurHash3;
