+++ java/fury-core/src/test/java/io/fury/builder/JITContextTest.java
@@ -22,9 +22,9 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.resolver.MetaContext;
-import io.fury.serializer.CompatibleMode;
 import io.fury.serializer.Serializer;
 import io.fury.test.bean.BeanA;
 import io.fury.test.bean.BeanB;
