+++ java/fury-core/src/test/java/io/fury/resolver/MetaContextTest.java
@@ -19,8 +19,8 @@ package io.fury.resolver;
 import com.google.common.collect.ImmutableList;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
-import io.fury.serializer.CompatibleMode;
 import io.fury.test.bean.BeanA;
 import io.fury.test.bean.BeanB;
 import io.fury.test.bean.Foo;
