+++ java/fury-core/src/test/java/io/fury/serializer/CompatibleSerializerTest.java
@@ -20,6 +20,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.test.bean.BeanA;
 import io.fury.test.bean.BeanB;
