+++ java/fury-core/src/test/java/io/fury/serializer/DuplicateFieldsTest.java
@@ -21,6 +21,7 @@ import static org.testng.Assert.assertEquals;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
 import io.fury.builder.CodecUtils;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.memory.MemoryBuffer;
 import io.fury.memory.MemoryUtils;
