+++ java/fury-core/src/test/java/io/fury/serializer/MetaSharedCompatibleTest.java
@@ -23,6 +23,7 @@ import com.google.common.collect.Sets;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
 import io.fury.builder.MetaSharedCodecBuilder;
+import io.fury.config.CompatibleMode;
 import io.fury.config.Language;
 import io.fury.resolver.MetaContext;
 import io.fury.test.bean.BeanA;
