+++ java/fury-core/src/test/java/io/fury/serializer/UnexistedClassSerializersTest.java
@@ -22,6 +22,7 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 import io.fury.Fury;
 import io.fury.FuryTestBase;
+import io.fury.config.CompatibleMode;
 import io.fury.config.FuryBuilder;
 import io.fury.config.Language;
 import io.fury.resolver.MetaContext;
